requirements.txt
langchain
langchain_openai
langgraph-cli
dotenv
deepagents
minio
markitdown[all]
pypandoc
mcp
langchain-mcp-adapters